Fix path check for default value of DEPLOYDIR
In case DEPLOYDIR is not defined in the `local.conf`, `find_files`
tries to use default `$BUILDDIR/tmp/deploy` and assigns this value to
`global_values.deploydir`.
But, when verifying if DEPLOYDIR exists, it uses temporary variable
`deploydir` reflecting value from the `local.conf`.
Obviously, when DEPLOYDIR was not defined in the `localconf`,
the temporary variable `deploydir` is an empty string.
As a result path check always fails in case DEPLOYDIR is not set in
the `local.conf`.